+++ src/ducks/payments/hooks/usePayment.js
@@ -1,9 +1,5 @@
-import { useCallback, useEffect, useState } from 'react'
-import {
-  createBiPayment,
-  createPaymentCreation,
-  getUrlWebView
-} from '../service'
+import { useEffect, useState } from 'react'
+import { createBiPayment, getUrlWebView } from '../service'
 import { usePaymentContext } from '../PaymentContext'
 import { useClient } from 'cozy-client'
 
@@ -29,41 +25,11 @@ export const useAccessToken = () => {
   return state
 }
 
-const usePaymentCreation = () => {
-  const [state, setState] = useState({})
-  const { payment, setPayment, biPayment } = usePaymentContext()
-  const client = useClient()
-
-  useEffect(() => {
-    const loadPaymentCreation = async () => {
-      setState(st => ({ ...st, status: 'loading' }))
-      try {
-        const paymentCreation = await createPaymentCreation({
-          client,
-          payment,
-          biPayment
-        })
-        setState(st => ({
-          ...st,
-          paymentCreation,
-          status: 'done',
-          error: null
-        }))
-        setPayment(paymentCreation)
-      } catch (e) {
-        setState(st => ({ ...st, status: 'error', error: e }))
-      }
-    }
-    loadPaymentCreation()
-  }, [biPayment, client, payment, setPayment])
-
-  return { ...state }
-}
 export const useUrlWebView = (paymentId, biPayment) => {
   const [webViewUrl, setWebViewUrl] = useState('')
   useEffect(() => {
     const loadUrlWebView = async () => {
-      if (payment.id) {
+      if (paymentId) {
         const url = await getUrlWebView(paymentId, biPayment)
         setWebViewUrl(url)
       }
